Update response and solve issue #4
1. This first updates the response returned by the server. Previously
all responses were wrapped in a response object. Now if you ask for a
task, a task is returned, not wrapped inside a response object.
2. Addressed issue number 4. When a task was added to a day with a task
name, there was an javascript error because when rendering the slot,
the newly created task was not available in the tasks. Now create a
slot with a task name, return the created task and the slot;