Do not show the form after task has been added
Currently: When adding a task to a day, the form sometimes still
appears even after the operation was successful hence giving the
impression that the operation failed.
Now: All remove state to the form component and pass all event and
actions handling to the enclosing EmptySlot component. So EmptySlot
listens to SlotCreated Event and update its state as not currently
editing and the form will not be shown